% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{covest-package}
\alias{covest-package}
\title{Nadaraya-Watson type kernel estimation of the conditional mean and covariance}
\description{
Contains functions to estimate the conditional mean and covariance using a 
Nadaraya-Watson kernel based estimator. This package contains the 
Nadaraya-Watson kernel-based estimator for the conditional mean and covariance 
presented in Yin et al. (2010) and Neumann et al. (2024) which can be used to 
condition the mean or covariance between system outputs on confounders, e.g. 
temperature.
}
\details{
Neumann et al. (2024).
 Confounder-adjusted Covariances of System Outputs and Applications to Structural Health Monitoring.
 \emph{PREPRINT}, \strong{XX}, pp. 1--26.

 Yin et al. (2010).
 Nonparametric Covariance Model. \emph{Statistica Sinica} \strong{20}, 469-–79.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/neumannLizzie/conditional-covariance}
  \item Report bugs at \url{https://github.com/neumannLizzie/conditional-covariance/issues}
}

}
\author{
\strong{Maintainer}: Lizzie Neumann \email{<neumannl@hsu-hh.de>} (\href{https://orcid.org/0000-0003-2256-1127}{ORCID})

Authors:
\itemize{
  \item Philipp Wittenberg \email{pwitten@hsu-hh.de} (\href{https://orcid.org/0000-0001-7151-8243}{ORCID})
}

}
