% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_est.R
\name{meanest}
\alias{meanest}
\title{Nadaraya-Watson type kernel estimation of the conditional mean}
\usage{
meanest(x, z, znew, h, mean_old, sumK_old)
}
\arguments{
\item{x}{Numeric Matrix. An nxp matrix containing the data (of, e.g., sensor
measurements or features) where the columns refer to p different sensors and
n is the number of observations.}

\item{z}{Numeric Vector. Containing the measured confounder (e.g., temperature).}

\item{znew}{Numeric. The value for which the conditional covariance is to be estimated.}

\item{h}{Numeric. The bandwidths, i.e. the smoothing parameter.}

\item{mean_old}{Numeric Vector. A 1xp vector, containing the conditional mean at value znew.
Only not zero if the data set is split and the conditional mean is estimated
and updated piece by piece or new data is added.}

\item{sumK_old}{Numeric. The sum of kernels. Only not zero if the data set is
split and the conditional mean is estimated and updated piece by piece or new data is added.}
}
\value{
Returns a matrix where the first column contains the conditional mean
at value znew and the first element in the second column is the sum of kernels (sumK).
}
\description{
Estimation of the conditional mean using a Nadaraya-Watson type kernel estimatior.
}
\details{
The used kernel is the Gaussian kernel which equals a normal density with
mean 0 and chosen bandwidth h.
K = dnorm(z - znew, 0, h, false)

The conditional mean is estimated as follows
mest = (K*transpose(x) + sumK_old*mean_old) / (sum(K) + sumK_old).
}
\examples{
\dontrun{
x <- cbind(c(1,3,5),c(3,2,6),c(4,1,5))
z <- c(5.1,3.1,4.3)
znew   <- 3.2
h <- 1.3
mean_old <- rep(0,3)
sumK_old <- 0

covest::meanest(x=x,z=z,znew=znew,h=h,mean_old=cest_mean_oldold, sumK_old=sumK_old)
}
}
\references{
Neumann et al. (2024).
 Confounder-adjusted Covariances of System Outputs and Applications to Structural Health Monitoring.
 \emph{PREPRINT}, \strong{XX}, pp. 1--26.

 Yin et al. (2010).
 Nonparametric Covariance Model. \emph{Statistica Sinica} \strong{20}, 469-–79.
}
\author{
Lizzie Neumann
}
