% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{covest-package}
\alias{covest-package}
\title{Variable Life Adjusted Display and Other Risk-Adjusted Quality Control Charts}
\description{
Contains functions to set up risk-adjusted quality control charts in health care. For the
variable life adjusted display (VLAD) proposed by Lovegrove et al. (1997)
<doi:10.1016/S0140-6736(97)06507-0> signaling rules derived in Wittenberg et al. (2018)
 <doi:10.1002/sim.7647> are implemented. Additionally, for the risk-adjusted cumulative sum
 chart based on log-likelihood ratio statistic introduced by Steiner et al. (2000)
 <doi:10.1093/biostatistics/1.4.441> average run length and control limits  can be computed with
  fast and accurate Markov chain approximations developed in Knoth et al. (2019)
  <doi:10.1002/sim.8104>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/neumannLizzie/conditional-covariance}
  \item Report bugs at \url{https://github.com/neumannLizzie/conditional-covariance/issues}
}

}
\author{
\strong{Maintainer}: Philipp Wittenberg \email{pwitten@hsu-hh.de} (\href{https://orcid.org/0000-0001-7151-8243}{ORCID})

Authors:
\itemize{
  \item Lizzie Neumann \email{<neumannl@hsu-hh.de>} (\href{https://orcid.org/0000-0003-2256-1127}{ORCID})
}

}
