% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_est.R
\name{covest}
\alias{covest}
\title{Nadaraya-Watson type kernel estimation of the conditional covariance}
\usage{
covest(x, z, znew, h, cest_old, sumK_old, mx)
}
\arguments{
\item{x}{Numeric Matrix. An nxp matrix containing the data (of, e.g., sensor 
measurements or features) where the columns refer to p different sensors and 
n is the number of observations.}

\item{z}{Numeric Vector. Containing the measured confounder (e.g., temperature).}

\item{znew}{Integer. The value for which the conditional covariance is to be estimated.}

\item{h}{Numeric. The bandwidths, i.e. the smoothing parameter.}

\item{cest_old}{Numeric Matrix. Randomization probability. If the CUSUM statistic is
equal to the threshold \code{h}, an control chart alarm is triggered with
probability \code{gamma}.}

\item{sumK_old}{Numeric. Switch for activating randomization in order to allow
continuous ARL control.}

\item{mx}{Numeric Matrix. Head start value as integer multiple of \code{1/m}.
Should be an element of \code{0:hm}.}
}
\value{
Returns a matrix with p rows and p+1 columns. 
The first row of the p+1 column contains the sum of the kernel (sumK).
The first pxp matrix is the covariance matrix at value znew.
}
\description{
Estimation of the conditional covariance using a Nadaraya-Watson 
type kernel estimatior.
}
\details{
TODO: Equations how the cond cov is implemented
}
\examples{
\dontrun{
x <- cbind(c(1,3,5),c(3,2,6),c(4,1,5))
z <- c(5.1,3.1,4.3)
znew   <- 3.2
h <- 1.3
cest_old <- matrix(0,ncol=3,nrow=3)
sumK_old <- 0
mx <- cbind(c(1.2,2.9,3.8),c(3.2,2.2,1.1))

covest::covest(x=x,z=z,znew=znew,h=h,cest_old=cest_old, sumK_old=sumK_old, mx=mx)
}
}
\references{
Neumann et al. (2024).
 Confounder-adjusted Covariances of System Outputs and Applications to Structural Health Monitoring.
 \emph{PREPRINT}, \strong{XX}, pp. 1--26.

 Yin et al. (2010).
 Nonparametric Covariance Model. \emph{Statistica Sinica} \strong{20}, 469-–79.
}
\author{
Lizzie Neumann
}
