% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_est.R
\name{covest}
\alias{covest}
\title{Zero-state ARL of discrete CUSUM charts}
\usage{
covest(x, z, znew, h, cest_old, sumK_old, mx)
}
\arguments{
\item{x}{Numeric Matrix. An \R function of a discrete univariate Cumulative
Distribution Function taking a numeric first argument and returning a
numeric vector of the same length. For example, \code{\link{pbinom}} for
the Binomial or \code{\link{ppois}} for the Poisson distribution. See
\link{Distributions} for other standard distributions.}

\item{z}{Numeric. Enumerator of rational approximation of reference value
\code{k}.}

\item{znew}{Integer. Enumerator of rational approximation of control limit
value \code{h}.}

\item{h}{Numeric. Denominator of rational approximation of reference value.}

\item{cest_old}{Numeric Matrix. Randomization probability. If the CUSUM statistic is
equal to the threshold \code{h}, an control chart alarm is triggered with
probability \code{gamma}.}

\item{sumK_old}{Numeric. Switch for activating randomization in order to allow
continuous ARL control.}

\item{mx}{Numeric Matrix. Head start value as integer multiple of \code{1/m}.
Should be an element of \code{0:hm}.}
}
\value{
Returns a single value which is the Average Run Length.
}
\description{
Compute the ARL of CUSUM.
}
\details{
TODO: Equations how the CUSUMs are implemented
}
\examples{
\dontrun{
A   <- 500
mu0 <- 1
m   <- 20
km1 <- 1.3*m
km2 <- 2.2*m
CDF <- function(q) ppois(q, lambda=mu0)
## calibrate Bi-CUSUM
res1 <- bicusum_discrete_crit(CDF, mu0, A, km1, km2, m)
## ARLs of single CUSUM
C1 <- cusum_discrete_arl(CDF, km1, res1$hm1, m, gamma=res1$g1, rando=TRUE)
C2 <- cusum_discrete_arl(CDF, km2, res1$hm2, m, gamma=res1$g2, rando=TRUE)
## ARL of Bi-CUSUM with randomization
CC <- bicusum_discrete_arl(CDF, km1, km2, res1$hm1, res1$hm2, m, rando=TRUE, g1=res1$g1, g2=res1$g2)
cbind(round(cbind(C1, C2, CC), 3))
## ARL of Bi-CUSUM without randomization
bicusum_discrete_arl(CDF, km1, km2, res1$hm1, res1$hm2, m)
}
}
\references{
Brook and Evans (1972).
 An approach to the probability distribution of CUSUM run length.
 \emph{Biometrika}, \strong{59}, pp. 539--549
}
\author{
Lizzie Neumann
}
