% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_est.R
\name{meanest}
\alias{meanest}
\title{Nadaraya-Watson type kernel estimation of the conditional mean}
\usage{
meanest(x, z, znew, h, mean_old, sumK_old)
}
\arguments{
\item{x}{Numreic Matrix An \R function of a discrete univariate Cumulative
Distribution Function taking a numeric first argument and returning a
numeric vector of the same length. For example, \code{\link{pbinom}} for
the Binomial or \code{\link{ppois}} for the Poisson distribution. See
\link{Distributions} for other standard distributions.}

\item{z}{Numeric Vector. Enumerator of rational approximation of reference value
\code{k}.}

\item{znew}{Numeric. Enumerator of rational approximation of control limit
value \code{h}.}

\item{h}{Numeric. Denominator of rational approximation of reference value.}

\item{mean_old}{Numeric. Randomization probability. If the CUSUM statistic is
equal to the threshold \code{h}, an control chart alarm is triggered with
probability \code{gamma}.}

\item{sumK_old}{Boolean. Switch for activating randomization in order to allow
continuous ARL control.}

\item{mx}{Integer. Head start value as integer multiple of \code{1/m}.
Should be an element of \code{0:hm}.}
}
\value{
Returns a single value which is the Average Run Length.
}
\description{
Estimation of the conditional mean.
}
\details{
TODO: Equations how the CUSUMs are implemented
}
\examples{
\dontrun{
x <- cbind(c(1,3,5),c(3,2,6),c(4,1,5))
z <- c(5.1,3.1,4.3)
znew   <- 3.2
h <- 1.3
mean_old <- rep(0,3)
sumK_old <- 0

covest::meanest(x=x,z=z,znew=znew,h=h,mean_old=cest_mean_oldold, sumK_old=sumK_old)
}
}
\references{
Neumann et al. (2024).
 Confounder-adjusted Covariances of System Outputs and Applications to Structural Health Monitoring.
 \emph{PREPRINT}, \strong{XX}, pp. 1--26.

 Yin et al. (2010).
 Nonparametric Covariance Model. \emph{Statistica Sinica} \strong{20}, 469-–79.
}
\author{
Lizzie Neumann
}
